+++ common/src/main/java/com/common/entity/Likes.java
@@ -1,0 +1,28 @@
+package com.common.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Builder
+@Getter
+@Table(name = "likes")
+@AllArgsConstructor
+@NoArgsConstructor
+public class Likes {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name="likes_id")
+    private Long likesId;
+
+    private String memberId;
+
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "fairytale_id")
+    private Fairytale fairytale;
+}
