+++ fairytale/src/main/java/fairytale/service/SonnetService.java
@@ -312,7 +312,7 @@ public class SonnetService {
 
 
             /* 디비에 저장 */
-            Fairytale myFairytale = FairyTaleConverter.toFairyTale(findMember, title, 0F, 0F, 0, genre);
+            Fairytale myFairytale = FairyTaleConverter.toFairyTale(findMember, title, 0F, 0F, 0, genre, 0L);
 
 
             fairyTaleRepository.save(myFairytale);
@@ -342,6 +342,7 @@ public class SonnetService {
                     .imageUrl(imageUrls)
                     .mp3Url(mp3Urls)
                     //.keywords(keywords)
+                    .likeCount(myresult.getLikeCount())
                     .build();
 
         } catch (Exception e) {
@@ -359,394 +360,6 @@ public class SonnetService {
                 .replace("\n", " ")   // 줄바꿈 처리
                 .replace("\r", "")      // 캐리지 리턴 제거
                 .replace("\t", " ");    // 탭을 공백으로 변환
-    }
-
-//========================================================================================================
-
-    /*
-    * 스트리밍 응답 + 프롬프트 이미지 분리
-    * */
-
-    public Object createPromptText(String body) {
-        String text = String.format(
-                "Create a prompt text in detail for stable diffusion model per 2 sheets based on [%s].\n" +
-                        "All images generated by stable diffusion model must be consistent style and main character in fairy tale has consistent appearance.\n" +
-                        //"Make the result as json format like \n" +
-                        "The response is always a **valid JSON object** like \n" +
-                        "\"\n" +
-                        "{\n" +
-                        "\"prompt\":\" \"\\[\n" +
-                        "{\"prompt\": \"...\"},\n" +
-                        "{\"prompt\": \"...\"},\n" +
-                        "\\]\" \\n\n" +
-                        "}\n" +
-                        "\", which each value of list should be in one line.\n" +
-                        "All responses are under 3600 tokens." +
-//                        "Do not use double quotes in any dialogues or direct quotes of pages." +
-                        "Do not say any words except json response",
-
-                body
-        );
-
-        Map<String, Object> messages = new HashMap<>();
-        messages.put("role", "user");
-
-        Map<String, Object> content = new HashMap<>();
-        content.put("type", "text");
-        content.put("text", text);
-
-        messages.put("content", List.of(content));
-
-        //log.info("messages: {}", messages);
-
-
-        Map<String, Object> payload = new HashMap<>();
-        payload.put("max_tokens", 3600);
-        payload.put("messages", List.of(messages));
-        payload.put("anthropic_version", "bedrock-2023-05-31");
-
-        String requestBody = null;
-        try {
-            requestBody = objectMapper.writeValueAsString(payload);
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException(e);
-        }
-
-        InvokeModelRequest request = InvokeModelRequest.builder()
-                .modelId(SONNET_MODEL_ID)
-                .contentType("application/json")
-                .body(SdkBytes.fromUtf8String(requestBody))
-                //.body(SdkBytes.fromByteArray(requestBody.getBytes(StandardCharsets.UTF_8)))
-                .build();
-
-
-        InvokeModelResponse response = bedrockRuntimeClientForClaude.invokeModel(request);
-        String responseBody = new String(response.body().asByteArray(), StandardCharsets.UTF_8);
-
-        JsonNode prompt;
-        try {
-            // JSON 파싱을 위한 ObjectMapper 생성
-            ObjectMapper objectMapper = new ObjectMapper();
-
-            String fixedJson = preprocessJsonByInvoke(responseBody);
-
-            log.info("fixedJson: {}", fixedJson);
-
-            JsonNode rootNode = objectMapper.readTree(fixedJson);
-
-            log.info("RootNode: {}", rootNode);
-
-
-            JsonNode pcontent = rootNode.get("content");
-            log.info("Jackson Content: {}", content);
-
-            JsonNode usage = rootNode.get("usage");
-            log.info("Jackson Content usage: {}", usage);
-
-            JsonNode ptext = pcontent.get(0).get("text");
-            log.info("text: {}", text);
-
-            prompt = ptext.get("prompt");
-            log.info("prompt: {}", prompt);
-        } catch (Exception e) {
-            e.printStackTrace();
-            throw new FairytaleHandler(ErrorStatus.FAIRYTALE_BAD_REQUEST);
-        }
-
-        return prompt;
-    }
-
-    @Transactional
-    public SseEmitter createFtWithStreaming(String memberId, String genre, String gender, String challenge) {
-        Member findMember = memberRepository.findById(memberId).orElseThrow(() -> new MemberHandler(ErrorStatus.MEMBER_NOT_FOUND));
-
-
-        SseEmitter emitter = new SseEmitter(60 * 1000L);
-
-        String text = String.format(
-                "Create a fairy tale for %s kid using these elements: Genre: %s and Challenge: %s. Write in an enchanting, classic fairy tale style and a satisfying ending. Include vivid descriptions and a moral lesson.\n" +
-                        "Totally 14 pages of amount.\n" +
-                        "Divide each page by \"\\n\\n\"" +
-                        //"Make the result as json format like \n" +
-                        "The response is always like \n" +
-                        "{\\n" +
-                        "제목:\"...\", \\n" +
-                        "내용: [\\n" +
-                        "1장: \"...\" \\n" +
-                        "]" +
-                        ", which and each page of 내용 should be in one line.\n" +
-//            "The body texts have SSML tags for polly service." +
-                        "All responses are under 3600 tokens." +
-                        "All responses in Korean except prompt." +
-                        "Do not use double quotes in any dialogues or direct quotes of pages." +
-                        "Do not say any words except response",
-
-                gender, genre, challenge
-        );
-
-        Map<String, Object> messages = new HashMap<>();
-        messages.put("role", "user");
-
-        Map<String, Object> content = new HashMap<>();
-        content.put("type","text");
-        content.put("text", text);
-
-        messages.put("content", List.of(content));
-
-
-        Map<String, Object> payload = new HashMap<>();
-        payload.put("max_tokens", 3600);
-        payload.put("messages", List.of(messages));
-        payload.put("anthropic_version", "bedrock-2023-05-31");
-
-        String requestBody = null;
-        try {
-            requestBody = objectMapper.writeValueAsString(payload);
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException(e);
-        }
-
-
-        // Bedrock API 요청 생성
-        InvokeModelWithResponseStreamRequest request = InvokeModelWithResponseStreamRequest.builder()
-                .modelId(SONNET_MODEL_ID)
-                .contentType("application/json")
-                .accept("application/json")
-                .body(SdkBytes.fromUtf8String(requestBody)) // 요청 본문
-                .build();
-
-        StringBuilder completeResponseTextBuffer = new StringBuilder();
-
-        // 스트리밍 응답 핸들러
-        InvokeModelWithResponseStreamResponseHandler responseHandler = InvokeModelWithResponseStreamResponseHandler.builder()
-                .subscriber(InvokeModelWithResponseStreamResponseHandler.Visitor.builder().onChunk(chunk -> {
-                            // Extract and print the text from the model's native response.
-//                            var response = new JSONObject(chunk.bytes().asUtf8String());
-                            //var text = new JSONPointer("/text").queryFrom(response);
-
-                            String result = chunk.bytes().asUtf8String();
-
-                            ObjectMapper objectMapper = new ObjectMapper();
-
-                            JsonNode rootNode = null;
-                            try {
-                                rootNode = objectMapper.readTree(result);
-
-
-                            } catch (JsonProcessingException e) {
-                                throw new RuntimeException(e);
-                            }
-
-                            //String type = rootNode.get("type").toString();
-
-
-                            if (rootNode.has("type") && "content_block_delta".equals(rootNode.get("type").asText())) {
-                                JsonNode delta = rootNode.get("delta");
-                                log.info("delta: {}", delta);
-
-                                String streamingText = delta.get("text").toString().replaceAll("^\"|\"$", "");
-
-                                // Append the text to the response text buffer.
-                                completeResponseTextBuffer.append(streamingText);
-
-                                try {
-                                    emitter.send(SseEmitter.event().name("message").data(streamingText));
-                                } catch (IOException e) {
-                                    emitter.completeWithError(e);
-                                }
-                                log.info("result : {}\n", result);
-                            }
-
-
-
-                        })
-                        .build())
-                .onComplete(() -> {
-                    try {
-                        emitter.send(SseEmitter.event().name("complete").data("스트리밍 완료"));
-
-                        //FairyTaleResponseDto.FairyTaleCreateDto sonnetResultByInvoke = getStreamingResult(findMember, String.valueOf(completeResponseTextBuffer), genre);
-
-                        //emitter.send(SseEmitter.event().name("FairytaleId").data(sonnetResultByInvoke));
-
-                        log.info("Total Streaming text: {}", completeResponseTextBuffer);
-
-                    } catch (IOException e) {
-                        log.info("스트리밍 중단 원인: {}", e.getMessage());
-                        emitter.completeWithError(e);
-                    }
-
-
-                    emitter.complete();   // 스트리밍 종료.
-
-                })
-                .onError(emitter::completeWithError)
-                .build();
-
-        BedrockRuntimeAsyncClient asyncClient = BedrockRuntimeAsyncClient.builder()
-                .credentialsProvider(DefaultCredentialsProvider.create())
-                .region(Region.US_EAST_1)
-                .build();
-
-        // Bedrock API 요청 실행
-        asyncClient.invokeModelWithResponseStream(request, responseHandler);
-
-
-        return emitter;
-        //return completeResponseTextBuffer.toString();
-    }
-
-
-
-    public FairyTaleResponseDto.FairyTaleCreateDto getStreamingResult(Member findMember, String streamingResult, String genre) throws JsonProcessingException {
-        String fixedJson = preprocessJsonByInvoke(streamingResult);
-        log.info("Total Streaming text: {}", fixedJson);
-
-        JsonNode rootNode = objectMapper.readTree(fixedJson);
-
-        log.info("RootNode: {}", rootNode);
-
-
-        String title = rootNode.get("title").asText();
-        log.info("title: {}", title);
-
-        JsonNode bodyNode = rootNode.get("body");
-        log.info("body: {}", bodyNode);
-
-//        JsonNode promptNode = rootNode.get("prompt");
-//        log.info("prompt: {}", promptNode);
-
-        TreeMap<String, String> sortedBody = new TreeMap<>(Comparator.comparingInt(k -> Integer.parseInt(k.substring(4))));
-
-        for (JsonNode p: bodyNode){
-            p.fields().forEachRemaining(entry -> {
-                        String output = Jsoup.parse(entry.getValue().asText()).text();
-                        sortedBody.put(entry.getKey(), output);
-                    }
-            );
-        }
-        log.info("sortedBody: {}", sortedBody);
-
-        // Body를 하나의 문자열로 합치기
-        StringBuilder bodyText = new StringBuilder();
-        //sortedBody.forEach((key, value) -> bodyText.append(value).append("\n"));
-
-        StringBuilder resultPage = new StringBuilder();
-
-        List<String> pages = new ArrayList<>(sortedBody.values());
-        TreeMap<String, String> resultBody = new TreeMap<>();
-
-
-        TreeMap<String, String> sortedPrompt = new TreeMap<>();
-//		promptNode.fields().forEachRemaining(entry -> sortedPrompt.put(entry.getKey(), entry.getValue().asText()));
-
-//        for (JsonNode item : promptNode) {
-//            item.fields().forEachRemaining(entry -> sortedPrompt.put(entry.getKey(), entry.getValue().asText()));
-//        }
-//        log.info("Prompt: {}", sortedPrompt);
-
-
-        int j = 0;
-
-        for (int i=0; i< pages.size(); i++){
-
-            resultPage.append(pages.get(i));
-
-            if (pages.size()%2==1 && i==pages.size()-1) {  // resultBody의 마지막 인덱스일 때
-                if (sortedPrompt.size()%2==0){  // resultBody는 홀수 개수, prompt는 짝수 개수
-                    //j++;
-                    String key = "page" + j;
-                    //resultBody.put(key, String.valueOf(resultPage.append(pages.get(i))));
-                    String lastPaget = resultBody.get(key) + " " + pages.get(i);
-
-                    resultBody.replace(key, lastPaget);
-                } else {  // resultBody는 홀수 개수, prompt는 홀수 개수
-                    j++;
-                    String key = "page" + j;
-                    resultBody.put(key, String.valueOf(resultPage));
-                }
-
-            } else if (i%2==1){
-                j++;
-                String key = "page" + j;
-                resultBody.put(key, String.valueOf(resultPage));
-
-                resultPage = new StringBuilder("");
-
-            }
-
-
-        }
-
-        log.info("mergedBody: {}", resultBody);
-
-
-
-        // 결과 출력
-        System.out.println("Title: " + title);
-        System.out.println("\nBody:\n" + bodyText);
-
-        /********************* 이미지, mp3 생성 *********************/
-        List<FairyTaleRequestDto.StablediffusionRequestDto> imageRequestDtos = new ArrayList<>();
-        List<FairyTaleRequestDto.PollyRequestDto> mp3RequestDtos = new ArrayList<>();
-
-        List<String> prompts = sortedPrompt.values().stream().toList();
-        for (int i=1; i<=sortedPrompt.size(); i++){
-
-            String file = title + "-" + i;
-            log.info("imageFileName: {}", file);
-            log.info("prompt{} : {}", i-1, prompts.get(i-1));
-            imageRequestDtos.add(FairyTaleConverter.toImageRequestDto(title, file, prompts.get(i-1)));
-        }
-
-        List<String> mp3s = resultBody.values().stream().toList();
-        for (int i=1; i<=resultBody.size(); i++){
-
-            String file = title + "-" + i;
-            log.info("mp3FileName: {}", file);
-            log.info("mp3{}: {}", i-1,mp3s.get(i-1));
-            mp3RequestDtos.add(FairyTaleConverter.toMp3RequestDto(title,file, mp3s.get(i-1)));
-        }
-
-
-
-//            resultBody.forEach((key, value) -> {
-//                int i = 0;
-//                String file = title + i;
-//                mp3RequestDtos.add(FairyTaleConverter.toMp3RequestDto(title,file, value));
-//            });
-
-        log.info("Async image request: {}", imageRequestDtos);
-
-        /************************ 디비에 저장 ************************/
-        Fairytale myFairytale = FairyTaleConverter.toFairyTale(findMember, title, 0F, 0F, 0, genre);
-
-
-        fairyTaleRepository.save(myFairytale);
-
-        List<FairyTaleResponseDto.StablediffusionResultDto> imageUrls = fairyTaleService.asyncImage(imageRequestDtos, myFairytale);
-        List<FairyTaleResponseDto.PollyResultDto> mp3Urls = fairyTaleService.asyncPolly(mp3RequestDtos, myFairytale);
-
-
-        //log.info("image urls: {}", imageUrls);
-
-        sortedBody.forEach((key, value) -> {
-            String page = key.substring(4);
-            Integer pageNumber = Integer.parseInt(page);
-            log.info("page: {}", page);
-            log.info("pageNumber: {}", pageNumber);
-
-            Body body = FairyTaleConverter.toBody(value, pageNumber, myFairytale);
-            bodyRepository.save(body);
-        });
-
-
-        FairyTaleResponseDto.FairyTaleCreateDto sonnetResultByInvoke = FairyTaleResponseDto.FairyTaleCreateDto.builder()
-                .fairytaleId(myFairytale.getFairytaleId())
-                .build();
-
-
-        return sonnetResultByInvoke;
     }
 
 
@@ -914,7 +527,7 @@ public class SonnetService {
             for (JsonNode item : promptNode) {
                 item.fields().forEachRemaining(entry -> sortedPrompt.put(entry.getKey(), entry.getValue().asText()));
             }
-            log.info("Prompt: {}", sortedPrompt);
+            log.info("sortedPrompt: {}", sortedPrompt);
 
 
             int j = 0;
@@ -988,7 +601,7 @@ public class SonnetService {
             log.info("Async image request: {}", imageRequestDtos);
 
             /************************ 디비에 저장 ************************/
-            Fairytale myFairytale = FairyTaleConverter.toFairyTale(findMember, title, 0F, 0F, 0, genre);
+            Fairytale myFairytale = FairyTaleConverter.toFairyTale(findMember, title, 0F, 0F, 0, genre, 0L);
 
 
             fairyTaleRepository.save(myFairytale);
@@ -1018,6 +631,7 @@ public class SonnetService {
                     .body(resultBody)
                     .imageUrl(imageUrls)
                     .mp3Url(mp3Urls)
+                    .likeCount(myFairytale.getLikeCount())
                     //.keywords(keywords)
                     .build();
